feat: Read version directly from pubspec.yaml
Adds `pubspec.yaml` to the assets to allow for direct reading of the app version at runtime.

This commit refactors the `getAppVersion` service to prioritize fetching the version string by parsing the `pubspec.yaml` file included in the app's assets. This method serves as a more reliable primary source for the version. The `package_info_plus` plugin is now used as a fallback mechanism, with a hardcoded version as the final resort if both previous methods fail.