docs: Update release process documentation
The documentation for the release process has been updated. The CI/CD pipeline is now triggered on every push to master or when a pull request is merged, instead of when a new Git tag is pushed. The section for creating and pushing a Git tag has been removed.

Additionally, a minor extraneous line under the `flutter gen-l10n` command has been removed.