feat: Add recording overlay and floating action button
This commit introduces a new audio recording feature.

A `RecordingOverlay` widget has been added to provide a user interface for recording audio for a hymn. This overlay includes controls for starting, pausing, stopping, and saving a recording, as well as a countdown timer and visual feedback.

A floating action button (FAB) has been added to the `HymnDetailScreen`. This FAB allows users to either start a new recording or show the existing recording overlay if a recording is already in progress.

Dependencies for `googleapis` and `uuid` have been updated in `pubspec.yaml` to support this new functionality.